Move button components out of components/common.js
